+++ cmd/config.go
@@ -12,9 +12,7 @@ import (
 
 	"github.com/amazeeio/lagoon-cli/internal/lagoon"
 	"github.com/amazeeio/lagoon-cli/internal/lagoon/client"
-	"github.com/amazeeio/lagoon-cli/internal/schema"
 	"github.com/amazeeio/lagoon-cli/pkg/output"
-	"github.com/hashicorp/go-version"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
@@ -251,15 +249,15 @@ var configLagoonVersionCmd = &cobra.Command{
 			viper.GetString("lagoons."+current+".version"),
 			lagoonCLIVersion,
 			debug)
-		lagoonVersion, err := getLagoonAPIVersion(lc)
+		lagoonVersion, err := lagoon.GetLagoonAPIVersion(context.TODO(), lc)
 		if err != nil {
 			return err
 		}
-		viper.Set("lagoons."+cmdLagoon+".version", lagoonVersion)
+		viper.Set("lagoons."+cmdLagoon+".version", lagoonVersion.LagoonVersion)
 		if err = viper.WriteConfig(); err != nil {
 			return fmt.Errorf("couldn't write config: %v", err)
 		}
-		fmt.Println(lagoonVersion)
+		fmt.Println(lagoonVersion.LagoonVersion)
 		return nil
 	},
 }
@@ -284,96 +282,4 @@ func init() {
 	configAddCmd.Flags().StringVarP(&lagoonKibana, "kibana", "k", "", "Lagoon Kibana URL (https://logs-db-ui-lagoon-master.ch.amazee.io)")
 	configFeatureSwitch.Flags().StringVarP(&updateCheck, "disable-update-check", "", "", "Enable or disable checking of updates (true/false)")
 	configFeatureSwitch.Flags().StringVarP(&projectDirectoryCheck, "disable-project-directory-check", "", "", "Enable or disable checking of local directory for Lagoon project (true/false)")
-}
-
-// get the version of the api, or fall back to schema checks
-func getLagoonAPIVersion(lc *client.Client) (string, error) {
-	// always start at v1.0.0
-	lagoonVersion := "v1.0.0"
-
-	// if we have the api version available, just use it
-	lagoonAPIVersion, err := lagoon.GetLagoonAPIVersion(context.TODO(), lc)
-	if err != nil {
-		if !strings.Contains(err.Error(), `Cannot query field "lagoonVersion" on type "Query"`) {
-			return lagoonVersion, err
-		}
-	}
-	if lagoonAPIVersion.LagoonVersion != "" {
-		lagoonVersion = lagoonAPIVersion.LagoonVersion
-		if isValidSemver(lagoonVersion) {
-			// if we get a valid version from the api, drop off here
-			return lagoonVersion, nil
-		}
-	}
-	// otherwise lets try to determine from schema changes
-	lagoonSchema, err := lagoon.GetLagoonSchema(context.TODO(), lc)
-	if err != nil {
-		return lagoonVersion, err
-	}
-	return determineLagoonVersion(lagoonVersion, *lagoonSchema)
-}
-
-// determine the version of the API based on the schema
-func determineLagoonVersion(lagoonVersion string, lagoonSchema schema.LagoonSchema) (string, error) {
-	var err error
-	for _, schemaType := range lagoonSchema.Types {
-		if schemaType.Name == "Mutation" {
-			for _, field := range schemaType.Fields {
-				if field.Name == "switchActiveStandby" {
-					lagoonVersion, err = greaterThanOrEqualVersion(lagoonVersion, "v1.4.0")
-					if err != nil {
-						return lagoonVersion, err
-					}
-				}
-			}
-		}
-		if schemaType.Name == "Query" {
-			for _, field := range schemaType.Fields {
-				if field.Name == "allGroups" {
-					lagoonVersion, err = greaterThanOrEqualVersion(lagoonVersion, "v1.1.0")
-					if err != nil {
-						return lagoonVersion, err
-					}
-				}
-				if field.Name == "me" {
-					lagoonVersion, err = greaterThanOrEqualVersion(lagoonVersion, "v1.3.0")
-					if err != nil {
-						return lagoonVersion, err
-					}
-				}
-			}
-		}
-		if schemaType.Name == "NotificationMicrosoftTeams" {
-			lagoonVersion, err = greaterThanOrEqualVersion(lagoonVersion, "v1.2.0")
-			if err != nil {
-				return lagoonVersion, err
-			}
-		}
-	}
-	return lagoonVersion, nil
-}
-
-// return the given or greater than version
-func greaterThanOrEqualVersion(a string, b string) (string, error) {
-	aVer, err := version.NewSemver(a)
-	if err != nil {
-		return a, err
-	}
-	bVer, err := version.NewSemver(b)
-	if err != nil {
-		return b, err
-	}
-	if aVer.GreaterThanOrEqual(bVer) {
-		return a, nil
-	}
-	return b, nil
-}
-
-// check if string is valid semver
-func isValidSemver(a string) bool {
-	_, err := version.NewSemver(a)
-	if err != nil {
-		return false
-	}
-	return true
 }
